Address Chris' comments, with the exception of the resource restructuring and determining the base directory.

+++ src/pattern/types.h
@@ -1,5 +1,5 @@
 #include <inttypes.h>
 #include "../n64/types.h"
 
-typedef uint16_t TConfig;
+typedef uint32_t TConfig;
 typedef u64 TFlipFunc(TConfig* configs);
